PB-239: Refactor EmailValidationField.
