+++ src/utils/components/EmailValidationField.vue
@@ -5,9 +5,13 @@ import { isValidEmail } from '@/utils/utils'
 
 const emits = defineEmits(['email-updated'])
 const props = defineProps({
-    isRequestPending: {
+    disabled: {
         type: Boolean,
         default: false,
+    },
+    label: {
+        type: String,
+        default: '',
     },
 })
 
@@ -27,11 +31,12 @@ watch(email, () => {
 
 <template>
     <div>
-        <span><slot name="header"></slot></span>
-        <div class="input-group has-validation">
+        <div class="form-group has-validation">
+            <label for="email">{{ $t(props.label) }}</label>
             <input
+                id="email"
                 v-model="email"
-                :disabled="props.isRequestPending"
+                :disabled="props.disabled"
                 :class="{ 'is-invalid': !userIsTypingEmail && !isEmailValid }"
                 type="email"
                 class="form-control"
